Added new method that's less strict about height updates
Now the size of the nodes does not change immediately when the tree's height increases.

The size of the nodes is determined by that level which contains two nodes that are adjacent to one another.

The size is calculated in such a way that any adjacent nodes on this level will exactly touch each other