Fixed a bug causing incorrect height of the p5 canvas
It seems like the #canvas-wrapper div's height is getting used to set the height of the p5 canvas BEFORE the CSS is applied to that div to determine its layout, hence the canvas gets created with wrong height and position.

The hacky fix was to use setTimeout to manually call windowResized to resize the canvas moments after it first gets created.