Finally fixed the AVL tree bug
When nodes of equal value would be inserted into the AVL tree, sometimes subtrees would disappear, and a single node would become the child of multiple nodes.

This was due to the fact that joinWithParent() was using the values of the nodes to join a node with its parent. This would indirectly force a node to ALWAYS become the right child of its parent, if it happened to be the same value as its parent. But this is not the case when rotations are allowed.

A node could also become the left child of its parent if it was the same value as its parent... This was causing a node to get assigned as the left AND right child of its parent at the same time.

Damn, this was a fun bug to track down.